relocated datarecord, added comments, cleaning up
