+++ src/cpt/DataRecord.java
@@ -1,8 +1,15 @@
-package charts;
+package cpt;
 
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
 
+
+/**
+ * @author Noah Lin
+ * Class responsible for handling data in the table
+ * Creates datarecord object.
+ * 
+ */
 public class DataRecord {
     
 
@@ -10,6 +17,12 @@ public class DataRecord {
     private StringProperty country;
     private StringProperty value;
 
+    /**
+     * Constructor method for datarecord class. Contains all 3 values for 3 columns. 
+     * @param y year
+     * @param c country
+     * @param v value 
+     */
     public DataRecord(String y, String c, String v) {
         this.year = new SimpleStringProperty(y);
         this.country = new SimpleStringProperty(c);
@@ -17,15 +30,18 @@ public class DataRecord {
     }
 
 
+    //getter method for the year 
     public StringProperty yearProperty() {
         return year;
     }
 
+    //getter method for the country 
     public StringProperty countryProperty() {
         return country;
     }
 
 
+    //getter method for the coefficient value. 
     public StringProperty valueProperty() {
         return value;
