+++ src/cpt/HomeTable.java
@@ -1,114 +1,114 @@
 package cpt;
 
 import javafx.application.Application;
-import javafx.beans.value.ChangeListener;
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
-import javafx.event.Event;
 import javafx.event.EventHandler;
-import javafx.geometry.Insets;
-import javafx.geometry.Pos;
-import javafx.scene.Parent;
 import javafx.scene.Scene;
-import javafx.scene.control.TableColumn;
-import javafx.scene.control.TableRow;
-import javafx.scene.control.TableView;
-import javafx.scene.control.ToolBar;
-import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.layout.GridPane;
-import javafx.scene.paint.Color;
 import javafx.stage.Stage;
-import java.io.*;
-import java.util.List;
-
-import javax.sound.sampled.DataLine;
-import javax.swing.Action;
-
-import javafx.scene.control.ToolBar;
-import javafx.scene.robot.*;
 
 import javafx.scene.control.Button;
-import javafx.scene.control.ChoiceBox;
-import javafx.scene.Group;
 
-import charts.DataRecord;
-import charts.MyLineChart;
 
-import java.util.ArrayList;
-import cpt.ReadCSV;
-import cpt.DataTable;;
-
+/**
+ * @author Noah Lin
+ * This is the main class file that launches all of the scenes and controls the menu buttons.
+ * This class file should be run in order to use the program. 
+ */
 public class HomeTable extends Application{
     
     GridPane grid = new GridPane();
 
 
  
-
+    /**
+     * This method is responsible for handling all buttons and launching all scenes. 
+     * It is called when the class file is run. 
+     * 
+     * 
+     * 
+     */
     @Override public void start(Stage primaryStage) throws Exception {
     
+        //creating scene objects to represent the static methods to draw each type of graph
         Scene tableScene = new Scene(DataTable.makeTable());
         Scene lineChart = new Scene(DataLineChart.makeLineChart());
         Scene barScene = new Scene(BarGraphChart.makeBarChart());
 
-        Button myButton =  DataTable.lineButton;
+        //object and event handling for the 
+        Button btnHomeLineGraph =  DataTable.lineButton;
+        Button homeButton = DataLineChart.homeButton;
+        Button bargraphButton = DataTable.barButton;
+        Button bargrapghButton2 = DataLineChart.barButton;
+        Button homeButton2 = BarGraphChart.homeButton;
+        Button lineButton2 = BarGraphChart.lineButton;
 
-        myButton.setOnAction(new EventHandler<ActionEvent>() {
+
+        //event handling for homepage line graph button
+        btnHomeLineGraph.setOnAction(new EventHandler<ActionEvent>() {
             @Override public void handle(ActionEvent e) {
-
+                
+                //sets the primarystage to the line chart scene 
                 primaryStage.setScene(lineChart);
-                primaryStage.setResizable(true);
+                primaryStage.show();
             }
         });
 
-        Button homeButton = DataLineChart.homeButton;
 
+        //event handling for homebutton from the line chart 
         homeButton.setOnAction(new EventHandler<ActionEvent>() {
             @Override public void handle(ActionEvent e) {
+
+                //sets the primarystage to the tablescene
                 primaryStage.setScene(tableScene);
                 primaryStage.show();
             }
         });
 
-        Button bargraphButton = DataTable.barButton;
-
+        //event handling for the bargraph button from homescreen 
         bargraphButton.setOnAction(new EventHandler<ActionEvent>() {
             @Override
             public void handle(ActionEvent e) {
+
+                //sets scene to bargraph
                 primaryStage.setScene(barScene);
                 primaryStage.show();
             }
         });
 
-        Button bargrapghButton2 = DataLineChart.barButton;
-
+        //event handling for bargraphbutton from the line chart 
         bargrapghButton2.setOnAction(new EventHandler<ActionEvent>() {
             @Override public void handle (ActionEvent e) {
+
+                //sets scene to bargraph
                 primaryStage.setScene(barScene);
                 primaryStage.show();
             }
         });
 
-
-        Button homeButton2 = BarGraphChart.homeButton;
+        //event handling for the homebutton from the bargraph
         homeButton2.setOnAction(new EventHandler<ActionEvent>() {
             @Override public void handle(ActionEvent e) {
+
+                //sets scene to the datatable
                 primaryStage.setScene(tableScene);
                 primaryStage.show();
             }
         });
 
-        Button lineButton2 = BarGraphChart.lineButton;
+        //event handling for the line graph button from the bar graph 
         lineButton2.setOnAction(new EventHandler<ActionEvent>() {
             @Override public void handle (ActionEvent e) {
+
+                //sets scene to the line chart 
                 primaryStage.setScene(lineChart);
                 primaryStage.show();
             }
         });
 
-        primaryStage.setScene(tableScene);
 
+        //sets the scene to display the data table by default when application is run 
+        primaryStage.setScene(tableScene);
         primaryStage.setResizable(true);
         primaryStage.show();
 
@@ -117,6 +117,8 @@ public class HomeTable extends Application{
     }
 
 
+
+    
     public static void main(String[] args) {
         launch(args);
     }
